Fix some subtitle placement issues due to timing.
For bighalo subtitles, there was a typo that caused timing of the third subtitle text to be placed immediately. There were some math issues that caused the 4th subtitle timing to be just a bit off, causing breif overlap. Another user variable was added for a timing gap between 3rd and 4th subtitles for use with the 4 subtitles option.